transformation: fixed min value analaysis & offset return type
- negative slack bounds & negative/positive bias cases were not
  respected
- return type for offeset as usize makes negative offsets set to 0

+++ crates/python/src/transform/builtin/transformation/integer_to_binary.rs
@@ -21,7 +21,7 @@ impl PyArtifact {
     }
 
     #[getter]
-    fn offsetmap(&self) -> HashMap<String, usize> {
+    fn offsetmap(&self) -> HashMap<String, i64> {
         self.0.offsetmap().clone()
     }
